refactor: isolate recovery scheduling and reconciliation
